==> 1. Pre Processing Data.ipynb <==
Program untuk memisahkan beban listrik non hari libur nasional yang akan digunakan


Library :

import pandas as pd
import numpy as np
from National_Holiday import national_holiday
from Remove_Duplicate import Remove

Import Data :

path = r'kalbar1619.xlsx'
df01 = pd.read_excel (path)
df01.head()

df01.tail()

Membuat Kolom Waktu :

df01['year'] = df01['datetime'].apply(lambda x: x.year)
df01['quarter'] = df01['datetime'].apply(lambda x: x.quarter)
df01['month'] = df01['datetime'].apply(lambda x: x.month)
df01['day'] = df01['datetime'].apply(lambda x: x.day)
df01["weekday"]=df01.apply(lambda row:row["datetime"].weekday(),axis=1)
df01["weekday"] = (df01["weekday"] < 5).astype(int)
df01["DayNumber"]=df01['datetime'].apply(lambda x: x.weekday())

df01.head(1)

DayNumber data awal adalah 4 (seharusnya 0), maka data awal akan dipotong

df01=df01.drop(df01.index[0:(((24*3)-1))])
df01.head(2)

df01.tail(2)

DayNumber data akhir adalah 1 (Seharusnya 6), maka data akhir juga dipotong

df01=df01.drop(df01.index[-24-1:])
df01.tail(2)

df01 = df01.reset_index(drop=True)

Menandai Hari Libur Nasional 

data = df01.reset_index()
temp1 = len(df01)
temp2=[]
for i in range(temp1):
    temp3 = data.datetime.iloc[i]
    hari  = temp3.day
    bulan = temp3.month
    tahun = temp3.year
    weekday = data.weekday.iloc[i]
    temp4 = national_holiday(hari,bulan,tahun,weekday)
    temp2.append(temp4)

df01["National_Holiday"] = temp2

df01.head()

df01["WeekNumber"] = 0

Membuat Nomer Minggu

a = 0
for i in range(len(df01)):
#     print(i)
    if ((df01.DayNumber.iloc[i] == 0) and (df01.datetime.iloc[i].hour == 0)) :
        a = a+1
        df01.WeekNumber.iloc[i] = a
    else :
        df01.WeekNumber.iloc[i] = a
df01.tail()

Mencari Minggu yang terdapat hari libur nasional 

Delete_Week = []
for i in range(len(df01)):
    NH = df01.National_Holiday.iloc[i]
    WN = df01.WeekNumber.iloc[i]
    if NH == 1:
        Delete_Week.append(WN)
        
Delete_Week = Remove(Delete_Week)

Menghapus Minggu yang terdapat Hari Libur Nasional

for i in range(len(Delete_Week)):
    DW = Delete_Week[i]
    df01 = df01[df01.WeekNumber != DW]
df01 = df01.reset_index(drop=True)

df01.tail()

a = 0
for i in range(len(df01)):
#     print(i)
    if ((df01.DayNumber.iloc[i] == 0) and (df01.datetime.iloc[i].hour == 0)) :
        a = a+1
        df01.WeekNumber.iloc[i] = a
    else :
        df01.WeekNumber.iloc[i] = a
df01 = df01.reset_index(drop=True)
df01.tail()

Terdapat Data yang rusak pada Minggu ke 37, panjang data bukan 168, melainkan 144

Week37 = df01[df01.WeekNumber == 37]
len(Week37)

df01 = df01[df01.WeekNumber != 37]

a = 0
for i in range(len(df01)):
#     print(i)
    if ((df01.DayNumber.iloc[i] == 0) and (df01.datetime.iloc[i].hour == 0)) :
        a = a+1
        df01.WeekNumber.iloc[i] = a
    else :
        df01.WeekNumber.iloc[i] = a
df01 = df01.reset_index(drop=True)
df01.tail()

df01 = df01.reset_index(drop=True)
df01.tail()

Power Dataset :

df01.power

# df01.power.to_excel("power_dataset.xlsx")  
# print("saved")
