==> 2. Create Parameter LSTM.ipynb <==
Program untuk membuat tabel parameter yang akan diuji

import pandas as pd
from pandas import ExcelWriter

### 1. Pengujian Timesteps LSTM

features = [504]
hidden_layer = [1,2,3,4]
hidden_neuron = [336]
iterasi = 30

df_parameter = pd.DataFrame(columns = ['dataset','feature','hidden_neuron','hidden_layer'])

j = 1
for i1 in range(len(features)) :
    for i2 in range(len(hidden_layer)) :
        for i3 in range(len(hidden_neuron)) :
            for i4 in range(iterasi) :
                if j == iterasi+1 :
                    j = 1
                    
                df_parameter = df_parameter.append({'dataset':j,'hidden_neuron':hidden_neuron[i3],'feature':features[i1],'hidden_layer':hidden_layer[i2]}, ignore_index=True)
                j = j + 1

df_parameter['RMSE'] = 0
df_parameter['MAPE'] = 0

df_parameter

writer = ExcelWriter('RNN_hidden_layer.xlsx')
df_parameter.to_excel(writer,'Sheet1')
writer.save()
print('complete')

### 2. Pengujian Features LSTM

*Tabel ini dibuat setelah melalui pengujian timesteps

features = [336, 504, 672]
hidden_layer = [1]
timesteps = [1]
iterasi = 1

df_parameter = pd.DataFrame(columns = ['timesteps','feature','hidden_layer'])

for i1 in range(len(features)) :
    for i2 in range(len(hidden_layer)) :
        for i3 in range(len(timesteps)) :
            for i4 in range(iterasi) :
                df_parameter = df_parameter.append({'timesteps':timesteps[i3],'feature':features[i1],'hidden_layer':hidden_layer[i2]}, ignore_index=True)

df_parameter['RMSE'] = 0
df_parameter['MAPE'] = 0

df_parameter

# writer = ExcelWriter('LSTM_features.xlsx')
# df_parameter.to_excel(writer,'Sheet1')
# writer.save()
# print('complete')

### 3. Pengujian Hidden Layer LSTM 

*Tabel ini dibuat setelah melakukan pengujian features

features = [336]
hidden_layer = [2,3]
timesteps = [1]
iterasi = 1

df_parameter = pd.DataFrame(columns = ['timesteps','feature','hidden_layer'])

for i1 in range(len(features)) :
    for i2 in range(len(hidden_layer)) :
        for i3 in range(len(timesteps)) :
            for i4 in range(iterasi) :
                df_parameter = df_parameter.append({'timesteps':timesteps[i3],'feature':features[i1],'hidden_layer':hidden_layer[i2]}, ignore_index=True)

df_parameter['RMSE'] = 0
df_parameter['MAPE'] = 0

df_parameter

# writer = ExcelWriter('LSTM_hidden_layer.xlsx')
# df_parameter.to_excel(writer,'Sheet1')
# writer.save()
# print('complete')
