==> 7. Perbandingan Model dan Pengujian Signifikansi (Koreksi).ipynb <==
import pandas as pd
import scipy
import numpy as np
from numpy import array

import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import mean_squared_error
from mape import mean_absolute_percentage_error

from IPython.display import clear_output
from pandas import ExcelWriter

import keras
from keras.models import Sequential
from keras.layers import LSTM
from keras.layers import SimpleRNN
from keras.layers import Dense
from keras.callbacks import EarlyStopping
from keras.models import model_from_json
from keras.layers import Dropout

from data_reshape import univariate_reshape

path = r'power_dataset.xlsx'
data = pd.read_excel (path)
target = data.power.iloc[-24*7:]
target = target.reset_index(drop=True)
dataset = data.power.iloc[:-24*7]
dataset = pd.DataFrame(dataset)

# Normalize Dataset )
scaler = MinMaxScaler(feature_range=(0, 1))
dataset = scaler.fit_transform(dataset)

# load LSTM MODEL
json_file = open('features parameter Revisi LSTM model number 0.json', 'r')
# json_file = open('LSTM5.json', 'r')
loaded_model_json = json_file.read()
json_file.close()
model_lstm = model_from_json(loaded_model_json)

# load weights into new model
model_lstm.load_weights("features parameter Revisi LSTM model number 0.h5")

# load RNN MODEL
json_file = open('features parameter Revisi RNN model number 0.json', 'r')
loaded_model_json = json_file.read()
json_file.close()
model_rnn = model_from_json(loaded_model_json)

# load weights into new model
model_rnn.load_weights("features parameter Revisi RNN model number 0.h5")

#LSTM 
in_steps_lstm = 1
in_features_lstm = 336
out_steps_lstm = 168
train_x,train_y,test_sample = univariate_reshape(dataset,in_steps_lstm,in_features_lstm,out_steps_lstm,target)

predict = model_lstm.predict(test_sample)
predict_lstm = scaler.inverse_transform(predict)

#RNN
in_steps_rnn = 1
in_features_rnn = 336
out_steps_rnn = 168
train_x,train_y,test_sample = univariate_reshape(dataset,in_steps_rnn,in_features_rnn,out_steps_rnn,target)

predict = model_rnn.predict(test_sample)
predict_rnn = scaler.inverse_transform(predict)

#ANN
in_steps_ann = 1
in_features_ann = 168
out_steps_ann = 168
train_x,train_y,test_sample = univariate_reshape(dataset,in_steps_ann,in_features_ann,out_steps_ann,target)
test_sample = test_sample.reshape(test_sample.shape[0],test_sample.shape[2])

predict = model_ann.predict(test_sample)
predict_ann = scaler.inverse_transform(predict)

j = 0

in_steps = 1
in_features = 168
out_steps = 168
    
# load ANN MODEL
json_file = open('hidden layer parameter Revisi ANN model number 0.json', 'r')
loaded_model_json = json_file.read()
json_file.close()
model_ann = model_from_json(loaded_model_json)

# load weights into new model
model_ann.load_weights("hidden layer parameter Revisi ANN model number 0.h5")
    
simpan_aktual = []
simpan_prediksi = []
simpan_rmse = []
simpan_mape = []
    
for i2 in range(30):
        
    print("Testing Model {}/{}\n".format(j,30))
    
    batas_bawah_target = 30 - i2
    batas_atas_target = batas_bawah_target-1
    if batas_atas_target == 0:
        target = data.power.iloc[-batas_bawah_target*24*7:]
    else :
        target = data.power.iloc[-batas_bawah_target*24*7:-batas_atas_target*24*7]
        
    target = target.reset_index(drop=True)
    dataset_trg = data.power.iloc[:-batas_bawah_target*24*7]
    dataset_trg = pd.DataFrame(dataset_trg)
        
    # Normalize Dataset
    scaler = MinMaxScaler(feature_range=(0, 1))
    dataset_trg = scaler.fit_transform(dataset_trg)
    _ , _ ,test_sample = univariate_reshape(dataset_trg,in_steps,in_features,out_steps,target)
    test_sample = test_sample.reshape(test_sample.shape[0],test_sample.shape[2])
        
    predict = model_ann.predict(test_sample)
    predict = scaler.inverse_transform(predict)
    
    RMSE = np.sqrt(mean_squared_error(target,predict[0]))
    MAPE = mean_absolute_percentage_error(target,predict[0])
    
    simpan_rmse.append(RMSE)
    simpan_mape.append(MAPE)
        
    simpan_aktual.append(target.values)
    simpan_prediksi.append(predict[0])
        
    if i2 == 29 :
        sp = np.array(simpan_prediksi)
        sp = sp.flatten()
        sa = np.array(simpan_aktual)
        sa = sa.flatten()

        simpan_total = pd.DataFrame(columns = ['Aktual','Prediksi'])
        simpan_total.Aktual = sa
        simpan_total.Prediksi = sp
            
    j = j+1
    clear_output(wait=False)
    
simpan_rmse_ANN = simpan_rmse
simpan_mape_ANN = simpan_mape

simpan_total

j = 0

in_steps = 1
in_features = 336
out_steps = 168
    
# load ANN MODEL
json_file = open('features parameter Revisi RNN model number 0.json', 'r')
loaded_model_json = json_file.read()
json_file.close()
model_rnn = model_from_json(loaded_model_json)

# load weights into new model
model_rnn.load_weights("features parameter Revisi RNN model number 0.h5")
    
simpan_aktual = []
simpan_prediksi = []
simpan_rmse = []
simpan_mape = []
    
for i2 in range(30):
        
    print("Testing Model {}/{}\n".format(j,30))
        
    batas_bawah_target = 30 - i2
    batas_atas_target = batas_bawah_target-1
    if batas_atas_target == 0:
        target = data.power.iloc[-batas_bawah_target*24*7:]
    else :
        target = data.power.iloc[-batas_bawah_target*24*7:-batas_atas_target*24*7]
    
    target = target.reset_index(drop=True)
    dataset_trg = data.power.iloc[:-batas_bawah_target*24*7]
    dataset_trg = pd.DataFrame(dataset_trg)
        
    # Normalize Dataset
    scaler = MinMaxScaler(feature_range=(0, 1))
    dataset_trg = scaler.fit_transform(dataset_trg)
    _ , _ ,test_sample = univariate_reshape(dataset_trg,in_steps,in_features,out_steps,target)
        
    predict = model_rnn.predict(test_sample)
    predict = scaler.inverse_transform(predict)
    
    RMSE = np.sqrt(mean_squared_error(target,predict[0]))
    MAPE = mean_absolute_percentage_error(target,predict[0])
    
    simpan_rmse.append(RMSE)
    simpan_mape.append(MAPE)
        
    simpan_aktual.append(target.values)
    simpan_prediksi.append(predict[0])
        
    if i2 == 29 :
        sp = np.array(simpan_prediksi)
        sp = sp.flatten()
        sa = np.array(simpan_aktual)
        sa = sa.flatten()

        simpan_total = pd.DataFrame(columns = ['Aktual','Prediksi'])
        simpan_total.Aktual = sa
        simpan_total.Prediksi = sp
            
    j = j+1
    clear_output(wait=False)
    
simpan_rmse_RNN = simpan_rmse
simpan_mape_RNN = simpan_mape

simpan_total

data.iloc[-1]

135-75

data.power.iloc[-168*16]

# Koreksi

#LSTM ANN
from scipy.stats import wilcoxon
stat, p = scipy.stats.ranksums(dfhasil.lstm, dfhasil.ann)
print('Statistics=%.3f, p=%.3f' % (stat, p))
# interpret
alpha = 0.05
if p > alpha:
    print('Sample is not significantly different (accept H0)')
else:
    print('Sample is significantly different (reject H0)')
    
print("")

#LSTM RNN
[stat,pval] = scipy.stats.ttest_ind(dfhasil.lstm,dfhasil.rnn)
alpha = 0.05;
print('T stat = {}'.format(stat))
print('P Value = {}'.format(pval))
if pval < alpha:
    print('Sample is significantly different (reject H0)')
else:
    print('Sample is not significantly different (accept H0)')

print("")

#LSTM ARIMA
[stat,pval] = scipy.stats.ttest_ind(dfhasil.lstm,dfhasil.arima)
alpha = 0.05;
print('T stat = {}'.format(stat))
print('P Value = {}'.format(pval))
if pval < alpha:
    print('Sample is significantly different (reject H0)')
else:
    print('Sample is not significantly different (accept H0)')
