==> 8. METODE PERAMALAN BEBAN DENGAN KOEFISIEN BEBAN.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

def LCF(dataset,load_day,load_hour) :
    
    #Mengumpulkan Beban pada Hari yang dituju
    target_day = dataset[dataset.DayNumber == load_day]
    target_day = target_day.reset_index(drop=True)

    day_count = len(target_day) / 24
#     print("Jumlah hari : {}".format(day_count))

    #Beban Hari target pada Jam target 
    day_target_hour = target_day[target_day.datetime.apply(lambda x: x.hour) == load_hour]
    day_target_hour = day_target_hour.reset_index(drop=True)
    day_target_hour = day_target_hour.power

    #Beban Maksimal pada hari setiap tanggalnya
    day_max = []
    for i in range(1,len(day_target_hour)+1):
        v1 = target_day[target_day.WeekNumber == i]
        v2 = max(v1.power)
        day_max.append(v2)

    #Prediksi dengan Menggunakan Metode Koefisien Beban : 
    #Mengihtung Alfa
    sum_alfa = 0
    sum_beta = 0
    for i in range(len(day_target_hour)):
        temp1 = day_target_hour[i]/day_max[i]
        sum_alfa = sum_alfa + temp1

        if i>= 1:
            temp2 = day_target_hour[i]
            temp3 = day_target_hour[i-1]
            sum_beta = sum_beta + (temp3/temp2) 

    alfa = sum_alfa / len(day_target_hour)
    beta = sum_beta / (len(day_target_hour)-1)
    hasil_prediksi = alfa*beta*day_max[-1]
    return hasil_prediksi

path = r'full_power_dataset_tanpa_libur.xlsx'
df = pd.read_excel (path)
df.head(1)

data_target = df.iloc[-30*168:]
data_target = data_target.reset_index(drop=True)
dataset = df.iloc[:-30*168]

#Prediksi 1 Minggu pada Dataset Uji
prediction = []
for i0 in range(30):
    print(i0)
    dataset = df.iloc[:-(30-i0)*168]
    hasil_peramalan_per_dataset = []
    for i1 in range(7):
        for i2 in range(24):
            koef_beban = LCF(dataset,i1,i2)
            hasil_peramalan_per_dataset.append(koef_beban)
    prediction.append(hasil_peramalan_per_dataset)

len(prediction)
